+++ stroom-core-shared/src/main/java/stroom/data/retention/shared/DataRetentionRulesResource.java
@@ -18,6 +18,7 @@ package stroom.data.retention.shared;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
 import org.fusesource.restygwt.client.DirectRestService;
 import stroom.util.shared.ResourcePaths;
 import stroom.util.shared.RestResource;
@@ -34,6 +35,7 @@ import javax.ws.rs.core.MediaType;
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public interface DataRetentionRulesResource extends RestResource, DirectRestService {
+
     @POST
     @Path("/read")
     @ApiOperation(
@@ -46,5 +48,5 @@ public interface DataRetentionRulesResource extends RestResource, DirectRestServ
     @ApiOperation(
             value = "Update data retention rules",
             response = DataRetentionRules.class)
-    DataRetentionRules update(DataRetentionRules dataRetentionRules);
+    DataRetentionRules update(@ApiParam("dataRetentionRules") DataRetentionRules dataRetentionRules);
 }