+++ stroom-core-shared/src/main/java/stroom/index/shared/IndexVolumeGroupResource.java
@@ -1,12 +1,14 @@
 package stroom.index.shared;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import org.fusesource.restygwt.client.DirectRestService;
 import stroom.entity.shared.ExpressionCriteria;
 import stroom.util.shared.ResourcePaths;
 import stroom.util.shared.RestResource;
 import stroom.util.shared.ResultPage;
+
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import org.fusesource.restygwt.client.DirectRestService;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
@@ -28,13 +30,13 @@ public interface IndexVolumeGroupResource extends RestResource, DirectRestServic
     @ApiOperation(
             value = "Finds index volume groups matching request",
             response = ResultPage.class)
-    ResultPage<IndexVolumeGroup> find(ExpressionCriteria request);
+    ResultPage<IndexVolumeGroup> find(@ApiParam("request") ExpressionCriteria request);
 
     @POST
     @ApiOperation(
             value = "Creates an index volume group",
             response = IndexVolumeGroup.class)
-    IndexVolumeGroup create(String name);
+    IndexVolumeGroup create(@ApiParam("name") String name);
 
     @GET
     @Path("/{id}")