+++ stroom-core-shared/src/main/java/stroom/pipeline/shared/TextConverterResource.java
@@ -18,6 +18,7 @@ package stroom.pipeline.shared;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
 import org.fusesource.restygwt.client.DirectRestService;
 import stroom.docref.DocRef;
 import stroom.util.shared.ResourcePaths;
@@ -35,17 +36,18 @@ import javax.ws.rs.core.MediaType;
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public interface TextConverterResource extends RestResource, DirectRestService {
+
     @POST
     @Path("/read")
     @ApiOperation(
             value = "Get a text converter doc",
             response = TextConverterDoc.class)
-    TextConverterDoc read(DocRef docRef);
+    TextConverterDoc read(@ApiParam("docRef") DocRef docRef);
 
     @PUT
     @Path("/update")
     @ApiOperation(
             value = "Update a text converter doc",
             response = TextConverterDoc.class)
-    TextConverterDoc update(TextConverterDoc xslt);
+    TextConverterDoc update(@ApiParam("xslt") TextConverterDoc xslt);
 }