+++ stroom-core-shared/src/main/java/stroom/pipeline/shared/XsltResource.java
@@ -18,6 +18,7 @@ package stroom.pipeline.shared;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
 import org.fusesource.restygwt.client.DirectRestService;
 import stroom.docref.DocRef;
 import stroom.util.shared.ResourcePaths;
@@ -37,12 +38,13 @@ import javax.ws.rs.core.MediaType;
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
 public interface XsltResource extends RestResource, DirectRestService {
+
     @POST
     @Path("/read")
     @ApiOperation(
             value = "Get an xslt doc",
             response = XsltDoc.class)
-    XsltDoc read(DocRef docRef);
+    XsltDoc read(@ApiParam("docRef") DocRef docRef);
 
     @PUT
     @Path("/update")
@@ -58,5 +60,5 @@ public interface XsltResource extends RestResource, DirectRestService {
     @POST
     @Path("/{xsltId}")
     void save(@PathParam("xsltId") final String xsltId,
-                         final XsltDTO xsltDto);
+              @ApiParam("xsltDto") final XsltDTO xsltDto);
 }